server/requirements.txt
#Python==2.7.10
tornado==4.3
PyYAML==3.12
SQLAlchemy==1.2.6
requests
python-consul
pycrypto
MySQL-python